requirements.txt
# Core dependencies
httpx>=0.27.0
web3>=6.15.0
eth-account>=0.11.0
py-clob-client>=0.17.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0
aiosqlite>=0.19.0
orjson>=3.9.0
typer>=0.12.0
rich>=13.7.0
pyyaml>=6.0.0
tweepy>=4.14.0
aiohttp>=3.9.0

# Web application dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0
python-multipart>=0.0.9
websockets>=13.0

# macOS Menu Bar App
rumps>=0.4.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-UserNotifications>=10.0
websocket-client>=1.7.0

# Dev dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
py2app>=0.28.0

# Testing dependencies for API tests
httpx>=0.27.0
